+++ gateway/src/apicast/configuration_loader/oidc.lua
@@ -21,8 +21,13 @@ _M.discovery = require('resty.oidc.discovery').new()
 
 local function load_service(service)
     if not service or not service.proxy then return nil end
+    local result = _M.discovery:call(service.proxy.oidc_issuer_endpoint)
 
-    return _M.discovery:call(service.proxy.oidc_issuer_endpoint)
+    if result and service.id then
+      result.service_id = service.id
+    end
+
+    return result
 end
 
 function _M.call(...)